+++ CHANGELOG.md
@@ -14,6 +14,7 @@ project adheres to
   cleanup. PR #97.
 * Update optional rsass dependency to 0.16.0.
 * Add optional support for tide 0.14 and 0.15.
+* Update gotham to 0.5 in example.
 
 
 ## Release 0.12.0 - 2020-08-14